Adds visiting unit type for optional data
When field is optional the QueryResult will end up visiting a unit type
instead of an option type (none or some). This is part of the default
implementation of serde (as opposed to serde_json). For some reason, the
nesting of the field `data_set` within it seems to trigger a fallback to
the original serde code and not the serde_json code.

To reproduce this, an optional field is added to the query result
finished test. This value is then set to null in the test data and
loaded in. Specifying that it visits a unit corrects the behavior by
mapping it to a None value.

Fixes #37